midterm_test_all_fixed/requirements.txt
langchain
langchain-google-genai
langchain-community
python-dotenv